Fixed a couple of issues. Dirty bit throwing alert when clicking some buttons, reset when saving changes.

+++ src/components/Items/NoteItem.js
@@ -8,70 +8,28 @@ import Clear from '@material-ui/icons/Clear';
 import 'bootstrap/dist/css/bootstrap.css';
 import ListGroup from 'react-bootstrap/ListGroup';
 
+
+
 import CaptureContext from '../../capture/capture-context';
 import UnsavedContext from '../../capture/unsaved-context';
 
-import { Form, Button, Alert, Row, Col, Modal} from 'react-bootstrap/';
-
-
 // import NoteFormDirtyAlert from '../UI/Notes/NoteFormDirtyAlert';
-
-function NoteFormDirtyAlert() {
-    const [show, setShow] = useState(true);
-  
-    const handleClose = () => setShow(false);
-    const handleShow = () => setShow(true);
-  
-    return (
-      <>
-        {/* <Button variant="primary" onClick={handleShow}>
-          Launch demo modal
-        </Button> */}
-  
-        <Modal show={show} onHide={handleClose}>
-          <Modal.Header closeButton>
-            <Modal.Title>Unsaved Changes</Modal.Title>
-          </Modal.Header>
-          <Modal.Body>You have unsaved changes. Click on Ok to proceed. Click on Cancel to resume editing.</Modal.Body>
-          <Modal.Footer>
-            <Button variant="secondary" onClick={handleClose}>
-              Cancel
-            </Button>
-            <Button variant="primary" onClick={handleClose}>
-              Ok
-            </Button>
-          </Modal.Footer>
-        </Modal>
-      </>
-    );
-}
 
 
 
 const NoteItem = (props) => {
 
-    //note alert
-
-    const [show, setShow] = useState(true);
-  
-    const handleClose = () => setShow(false);
-    const handleShow = () => setShow(true);
-
     const notesCtx = useContext(CaptureContext);
     const flagCtx = useContext(UnsavedContext);
 
     const [style, setStyle] = useState({display: 'none'});
-    const [loadAlert, setLoadAlert] = useState(false);
+  
 
-    // note alert
-
-    const resetAlert = () => {
-        setLoadAlert(false);
-    }
 
     const toggleSelect = () => {
         if(flagCtx.flag){
-            handleShow();
+            
+            flagCtx.setShow(true);
             
         } else {
             notesCtx.selectedItem(props.listItem);
@@ -83,7 +41,12 @@ const NoteItem = (props) => {
     };
 
     const deleteItem = () => {
-        props.deleteItem(props.listItem.id);
+        if(flagCtx.flag) {
+          flagCtx.setShow(true);
+        } else {
+          props.deleteItem(props.listItem.id);
+        }
+        
     }
 
     return(
@@ -95,20 +58,6 @@ const NoteItem = (props) => {
         }}>
             <ListGroup.Item className={props.selectStyle}>
             
-            <Modal show={show} onHide={handleClose}>
-            <Modal.Header closeButton>
-                <Modal.Title>Unsaved Changes</Modal.Title>
-            </Modal.Header>
-            <Modal.Body>You have unsaved changes. Click on Ok to proceed. Click on Cancel to resume editing.</Modal.Body>
-            <Modal.Footer>
-                <Button variant="secondary" onClick={handleClose}>
-                Cancel
-                </Button>
-                <Button variant="primary" onClick={handleClose}>
-                Ok
-                </Button>
-            </Modal.Footer>
-            </Modal>
             
             <span  onClick={toggleSelect}>{props.listItem.title}</span><Clear style={style}  onClick={deleteItem}/></ListGroup.Item>
         </ListGroup>
